+++ app/Http/Controllers/Backend/JobController.php
@@ -15,6 +15,8 @@ class JobController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', Job::class);
+
         $companyId = Auth::user()->company_id;
         $jobsData = Job::where('company_id', $companyId)->paginate(10);
         return Inertia::render('Backend/Jobs/Index', [
@@ -28,6 +30,8 @@ class JobController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Job::class);
+
         return Inertia::render('Backend/Jobs/Edit');
     }
 
@@ -36,6 +40,8 @@ class JobController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Job::class);
+
         $request->validate([
             'title' => 'required|string|max:255',
             'location' => 'required|string|max:255',
@@ -74,6 +80,8 @@ class JobController extends Controller
      */
     public function show(Job $job)
     {
+        $this->authorize('view', Job::class);
+
         $jobData = Auth::user()->company->jobs()->with('user')->find($job->id);
         return Inertia::render('Backend/Jobs/Show', [
             'jobData' => $jobData
@@ -85,6 +93,8 @@ class JobController extends Controller
      */
     public function edit(Job $job)
     {
+        $this->authorize('update', Job::class);
+
         $jobData = Auth::user()->company->jobs()->find($job->id);
         return Inertia::render('Backend/Jobs/Edit', [
             'jobData' => $jobData
@@ -96,6 +106,8 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job)
     {
+        $this->authorize('update', Job::class);
+
         $validated = $request->validate([
             'title' => 'required|string|max:255',
             'location' => 'required|string|max:255',
@@ -123,6 +135,8 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        $this->authorize('delete', Job::class);
+
         $job->delete();
         return redirect()->back()->with('success', 'Job deleted successfully');
     }
